requirements.txt
asgiref==3.4.1
Django==3.2.9
django-mathfilters==1.0.0
mysqlclient==2.0.3
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
